TIP-384: rename the translated label provider
