+++ src/Pim/Bundle/ImportExportBundle/Resources/config/job_labels.yml
@@ -1,0 +1,9 @@
+parameters:
+    pim_import_export.job_label.translated_label_provider.class: Pim\Bundle\ImportExportBundle\JobLabel\TranslatedLabelProvider
+
+services:
+    pim_import_export.job_label.translated_label_provider:
+        class: %pim_import_export.job_label.translated_label_provider.class%
+        arguments:
+            - '@translator'
+            - 'batch_jobs'
