+++ src/Pim/Bundle/ImportExportBundle/Resources/config/normalizers.yml
@@ -7,7 +7,7 @@ services:
         class: %pim_import_export.normalizer.job_execution.class%
         arguments:
             - '@translator'
-            - '@pim_import_export.provider.job_label'
+            - '@pim_import_export.job_label.translated_label_provider'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
@@ -17,7 +17,7 @@ services:
         arguments:
             - '@translator'
             - '@pim_catalog.localization.presenter.datetime'
-            - '@pim_import_export.provider.job_label'
+            - '@pim_import_export.job_label.translated_label_provider'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
